Move functor to its own file to depend on op
